# Backend/microservices/api.gateway/docs/swagger.yaml
basePath: /api
definitions:
...
  domain.ApiCompanyColorPalette:
    properties:
      accent_color:
        type: string
      company_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      primary_color:
        type: string
      secondary_color:
        type: string
    type: object
...
  domain.CreateCompanyColorPaletteRequest:
    properties:
      accent_color:
        type: string
      company_id:
        type: integer
      name:
        type: string
      primary_color:
        type: string
      secondary_color:
        type: string
    required:
    - company_id
    - name
    - primary_color
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta: {}
      statusCode:
        type: integer
      success:
        type: boolean
      validationErrors:
        items:
          type: string
        type: array
    type: object
host: localhost:4000
info:
  contact: {}
  description: Gateway unificado de entrada do ERP Lumini Hub, gerenciando roteamento
    de microsserviços.
  title: Lumini Hub ERP - API Gateway
  version: "1.0"
paths:
...
  /company-color-palettes:
    post:
      consumes:
      - application/json
      description: Cria uma nova paleta de cores personalizada (nome + 3 cores) pra
        uma empresa
      parameters:
      - description: Dados da Paleta
        in: body
        name: palette
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCompanyColorPaletteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ApiCompanyColorPalette'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Salva paleta de cores personalizada
      tags:
      - CompanyColorPalette
  /company-color-palettes/{id}:
    delete:
      consumes:
      - application/json
      description: Remove uma paleta de cores personalizada pelo ID
      parameters:
      - description: ID da Paleta
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Exclui paleta de cores personalizada
      tags:
      - CompanyColorPalette
  /company-color-palettes/by-company/{companyId}:
    get:
      consumes:
      - application/json
      description: Retorna as paletas de cores salvas pelo usuário pra uma empresa,
        mais recentes primeiro
      parameters:
      - description: ID da Empresa
        in: path
        name: companyId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ApiCompanyColorPalette'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Lista paletas de cores personalizadas por empresa
      tags:
      - CompanyColorPalette
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"